% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_tags.R
\name{fc_list}
\alias{fc_list}
\title{List Elements in HTML}
\usage{
fc_list(title, ..., .hsize = "6", .ltype = "ul")
}
\arguments{
\item{title}{(character) title for flash card}

\item{...}{(r ellipsis) unordered elements}

\item{.hsize}{(integer) specify header size (1-6)}

\item{.ltype}{(character) specify list type (ol/ul)}
}
\value{
Shiny Tags
}
\description{
List Elements in HTML
}
\examples{
\dontrun{
  fc_list(title = "This is the Title", "1", "2")
}
}
